+++ ydb/core/persqueue/pqtablet/partition/mlp/ut/ya.make
@@ -13,6 +13,7 @@ SRCS(
     mlp_child_sync_ut.cpp
     mlp_commit_ut.cpp
     mlp_consumer_order_ut.cpp
+    mlp_consumer_precharge_fifo_ut.cpp
     mlp_consumer_split_ut.cpp
     mlp_consumer_ut.cpp
     mlp_consumer_wakeup_ut.cpp
